Change writing by chatId to userId

+++ src/main/java/com/train4game/munoon/models/Key.java
@@ -12,7 +12,7 @@ public class Key {
 
     private boolean used;
 
-    private Long usedBy;
+    private int usedBy;
 
     private Player votedFor;
